6.93

+++ src/main/java/jp/co/sony/ppog/entity/City.java
@@ -34,7 +34,7 @@ public final class City implements Serializable {
 	 * This field corresponds to the database column ID
 	 */
 	@Id
-	private Long id;
+	private Integer id;
 
 	/**
 	 * This field corresponds to the database column NAME
@@ -58,7 +58,7 @@ public final class City implements Serializable {
 	 * This field corresponds to the database column POPULATION
 	 */
 	@Column(nullable = false)
-	private Long population;
+	private Integer population;
 
 	/**
 	 * This field corresponds to the database column LOGIC_DELETE_FLG
